move file to node_builder
